naver_stock/requirements.txt
# Web scraping
requests==2.31.0
beautifulsoup4==4.12.3
lxml==5.1.0
aiohttp==3.9.1

# Database
psycopg2-binary==2.9.9

# Data processing
pandas==2.2.0

# Cloud storage and BigQuery
google-cloud-storage==2.14.0
google-cloud-bigquery==3.14.0
pyarrow==15.0.0

# Environment management
python-dotenv==1.0.1
